@@ -1,32 +1,18 @@
-[build-system]
-requires = [
-    "setuptools>=75.8.0",
-    "wheel",
-    "build",
-]
-build-backend = "setuptools.build_meta"
-
 [project]
 name = "kenobix"
-dynamic = [
-    "optional-dependencies",
-]
 version = "5.0.0"
-requires-python = ">=3.9"
-dependencies = [
-    "cattrs>=25.3.0",
-]
-description = "High-performance document database with SQLite3 JSON optimization (15-665x faster than basic implementations). Based on KenobiDB by Harrison Erd."
+requires-python = ">=3.11"
+description = "Simple document database with SQLite3 JSON optimization, based on KenobiDB by Harrison Erd."
 readme = "README.md"
 
 classifiers = [
     "Environment :: Console",
     "Development Status :: 4 - Beta",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
     "License :: OSI Approved :: BSD License",
     "Intended Audience :: Developers",
     "Topic :: Database",
@@ -39,27 +25,45 @@ license = {file = "LICENSE"}
 
 authors = [
     {name = "Harrison Erd", email = "harrisonerd@gmail.com"},
+    {name = "Stefane Fermigier", email = "sf@fermigier.com"},
 ]
 
 maintainers = [
     {name = "KenobiX Contributors"},
 ]
 
+dependencies = [
+    "cattrs>=25.3.0",
+]
+
+[dependency-groups]
+dev = [
+    "pytest>=8.3.4",
+    "ruff>=0.14.0",
+    "pyrefly>=0.36.2",
+    "nox>=2025.5.1",
+    "coverage>=7.10.7",
+    "pytest-coverage>=0.0",
+]
+
 [project.urls]
-"Homepage" = "https://github.com/yourusername/kenobix"
-"Source Code" = "https://github.com/yourusername/kenobix"
-"Issue Tracker" = "https://github.com/yourusername/kenobix/issues"
+"Homepage" = "https://github.com/abilian/kenobix"
+"Source Code" = "https://github.com/abilian/kenobix"
+"Issue Tracker" = "https://github.com/abilian/kenobix/issues"
 "Original Project" = "https://github.com/patx/kenobi"
 
-[tool.setuptools.packages.find]
-where = ["src"]
-include = ["kenobix*"]
+[build-system]
+requires = ["uv-build>=0.8.4,<0.9.0"]
+build-backend = "uv_build"
 
-[tool.setuptools.dynamic]
-optional-dependencies.pip = { file = ['requirements/pip.lock'] }
-optional-dependencies.dev = { file = ['requirements/dev.lock'] }
-optional-dependencies.manage = { file = ['requirements/manage.lock'] }
+[tool.uv.build-backend]
+module-name = "kenobix"
+module-root = "src"
 
+
+#
+# Pytest
+#
 [tool.pytest.ini_options]
 markers = [
     "slow: marks tests as slow (deselect with '-m \"not slow\"')",
@@ -83,12 +87,3 @@ skip_covered = true
 skip_empty = true
 fail_under = 88
 show_missing = true
-
-[dependency-groups]
-dev = [
-    "pytest>=8.3.4",
-    "cattrs>=23.1.0",
-    "ruff>=0.14.0",
-    "pyrefly>=0.36.2",
-    "nox>=2025.5.1",
-]
