@@ -10,7 +10,6 @@ Requirements:
 """
 
 from dataclasses import dataclass
-from typing import List
 
 from kenobix import Document, KenobiX
 
@@ -33,7 +32,7 @@ class Post(Document):
     title: str
     content: str
     author_id: int
-    tags: List[str]
+    tags: list[str]
     published: bool = False
 
 
